sửa customer thành user, thêm UserRepository, UserService và UserServiceImpl
