--- src/main/java/com/codegym/controller/customer/CustomerController.java
@@ -1,4 +1,0 @@
-package com.codegym.controller.customer;
-
-public class CustomerController {
-}
