+++ src/main/java/com/codegym/controller/home/HomeController.java
@@ -1,4 +1,4 @@
-package com.codegym.controller;
+package com.codegym.controller.home;
 
 import com.codegym.model.Category;
 import com.codegym.model.Product;
