+++ src/main/java/com/codegym/model/OrderDetail.java
@@ -1,7 +1,6 @@
 package com.codegym.model;
 
 import javax.persistence.*;
-import java.util.List;
 
 @Entity
 @Table(name = "details")
