+++ src/main/java/com/codegym/model/User.java
@@ -1,38 +1,51 @@
 package com.codegym.model;
 
-
-
 import javax.persistence.*;
+import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
+import java.util.List;
 
 @Entity
-@Table(name = "users")
+@Table(name = "customers")
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long userId;
+    private Long customerId;
+
+    @NotEmpty
+    private String firstName;
+
+    @NotEmpty
+    private String lastName;
 
     @NotEmpty
     private String userName;
 
     @NotEmpty
+    @Email
+    private String email;
+
+    @NotEmpty
     private String password;
 
-    public User() {
-    }
+    @NotEmpty
+    private Long phoneNumber;
 
-    public User(Long userId, String userName, String password) {
-        this.userId = userId;
+    @NotEmpty
+    private String address;
+
+    public User(Long customerId, @NotEmpty String firstName, @NotEmpty String lastName, @NotEmpty String userName, @NotEmpty @Email String email, @NotEmpty String password, @NotEmpty Long phoneNumber, @NotEmpty String address) {
+        this.customerId = customerId;
+        this.firstName = firstName;
+        this.lastName = lastName;
         this.userName = userName;
+        this.email = email;
         this.password = password;
+        this.phoneNumber = phoneNumber;
+        this.address = address;
     }
 
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
+    public User() {
     }
 
     public String getUserName() {
@@ -43,6 +56,38 @@ public class User {
         this.userName = userName;
     }
 
+    public Long getCustomerId() {
+        return customerId;
+    }
+
+    public void setCustomerId(Long customerId) {
+        this.customerId = customerId;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
     public String getPassword() {
         return password;
     }
@@ -50,4 +95,21 @@ public class User {
     public void setPassword(String password) {
         this.password = password;
     }
+
+    public Long getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(Long phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
 }
