--- src/main/java/com/codegym/repository/CustomerRepository.java
@@ -1,4 +1,0 @@
-package com.codegym.repository;
-
-public interface CustomerRepository {
-}
