+++ src/main/java/com/codegym/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.codegym.repository;
+
+import com.codegym.model.User;
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+public interface UserRepository extends PagingAndSortingRepository<User,Long> {
+}
