--- src/main/java/com/codegym/service/CustomerService.java
@@ -1,4 +1,0 @@
-package com.codegym.service;
-
-public interface CustomerService {
-}
