+++ src/main/java/com/codegym/service/impl/UserServiceImpl.java
@@ -1,0 +1,40 @@
+package com.codegym.service.impl;
+
+import com.codegym.model.User;
+import com.codegym.repository.UserRepository;
+import com.codegym.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class UserServiceImpl implements UserService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Override
+    public User findById(Long id) {
+        return userRepository.findOne(id);
+    }
+
+    @Override
+    public User findByUserName(String userName) {
+        return null ;
+    }
+
+    @Override
+    public Iterable<User> findAll() {
+        return userRepository.findAll();
+    }
+
+    @Override
+    public User save(User user) {
+        userRepository.save(user);
+        return user;
+    }
+
+    @Override
+    public User remove(Long userId) {
+        User user = userRepository.findOne(userId);
+        userRepository.delete(userId);
+        return user;
+    }
+}
