+++ src/main/webapp/WEB-INF/views/user/delete.html
