+++ src/main/webapp/WEB-INF/views/user/list.html
