+++ src/main/webapp/WEB-INF/views/user/register.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+<form action="#" th:action="@{/register}"
+      th:object="${user}" method="POST" novalidate="novalidate">
+    <table>
+        <tr>
+            <td><input type="text" th:field="${user.id}"/>id</td>
+        </tr>
+        <tr>
+            <td><input type="text" th:field="${user.name}"/>name</td>
+        </tr>
+        <tr>
+            <td><input type="text" th:field="${user.email}"/>email</td>
+        </tr>
+        <tr>
+            <td><input type="text" th:field="${user.password}"/>password</td>
+        </tr>
+    </table>
+    <button type="submit" >SAVE</button>
+
+</form>
+</body>
+</html>
