+++ src/main/webapp/WEB-INF/views/user/view.html
